+++ src/max_div/_core/solver/__init__.py
@@ -7,6 +7,7 @@ k: number of vectors to be selected
 m: number of (group) constraints imposed on the problem.
 """
 
+from ._constraint_penalty import ConstraintPenalty
 from ._duration import TargetDuration, TargetIterationCount, TargetTimeDuration, hours, iterations, minutes, seconds
 from ._presets import SolverPreset
 from ._score import Score